Whole table system was rebuilt. Working fine!

+++ src/main/java/com/saidashevar/ptgame/controller/request/HireHeroRequest.java
@@ -7,5 +7,5 @@ public class HireHeroRequest {
 	private String gameId;
     private String login;
     private int coordinateY;
-    private Long cardId;
+    private int cardId;
 }
